refactor: валидация параметров пагинации убрана из server
