+++ server/src/main/java/ru/practicum/shareit/common/Utils.java
@@ -3,7 +3,6 @@ package ru.practicum.shareit.common;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 
-import static java.util.Objects.isNull;
 import static java.util.Objects.nonNull;
 
 public class Utils {
@@ -18,25 +17,5 @@ public class Utils {
 
     public static Pageable createOffsetBasedPageRequest(Integer from, Integer size) {
         return createOffsetBasedPageRequest(from, size, Sort.unsorted());
-    }
-
-    public static void validatePageableParams(Integer from, Integer size) {
-        if (isNull(from) && isNull(size)) {
-            return;
-        }
-
-        if ((isNull(from) && nonNull(size)) || (nonNull(from) && isNull(size))) {
-            throw new ValidationException("Параметры from и size должны быть оба указаны");
-        }
-
-        if (from == 0 && size == 0) {
-            throw new ValidationException("Оба параметра не могут быть равны нулю");
-        }
-        if (from < 0) {
-            throw new ValidationException("Параметр from не может быть меньше нуля");
-        }
-        if (size <= 0) {
-            throw new ValidationException("Параметр size не может быть меньше или равным нулю");
-        }
     }
 }
